@@ -3,14 +3,14 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.2
-  image_picker:
-  firebase_ml_vision:
+  cupertino_icons: ^1.0.3
+  image_picker: ^0.8.2
+  google_ml_kit: ^0.6.0
 
 dev_dependencies:
   flutter_test:
